Reorganized existing classes
Rationale for select changes:
- *Artist 1- -1 multiplicity associations with Song and Album:* In other
music library applications (like iTunes), each song has an associated
artist, but albums themselves may be assigned a specific artist. The
Project Overview also specifies that each album must have an associated
artist.
- *class Location instead of class Room:* In the requirements, the term
"room" is never used - only "location".
- *HAS:* Having a large many-to-many relationship-like class didn't seem
like good practice - it seemed like it would make managing relationships
between classes much more difficult once we began coding. I tried to
reorganize that code to make up for this.
- *pos int:* Although I suggested we store the pos in an intermediary
class, I now feel that in practice, it would be impractical to edit the
pos of each intermediary class to make any deletions/edits. When
implementing, we should add a method to the controller class that infers
the position of a song in a playlist by taking that song's position in the
song list of the playlist (which Umple will automatically generate).