Recreated HAS class as HomeAudioSystem
In commit 8daba47, I thought the HAS class meant "has" (denoting a
relationship) instead of "Home Audio System" (the system as a whole). I
re-added this HAS class with a new name (to avoid later confusion), and
created containment associations between it and other relevant
classes (as demonstrated in lecture).