Locations only have songs, fixed containment
Instead of allowing Playlists and Songs to be added to each Location,
decided to only allow Songs. When adding a Playlist to a room, we'll
just add all the playlist's song to a Location. This should make
implementation easier.

Previously, not all classes were contained inside HomeAudioSystem, but
now they are. Song is contained in AlbumTracklist because each Song must
belong to an Album.