One-to-many artists may be attached to one song
Although prof. suggested a zero-to-many relationship, to me it seems too vague to have only artists not associated with the album to be associated with the song. Also, if an album artist is a collective or just "Various Artists", we wouldn't necessarily want all songs to have "Various Artists" or the broader collective name as its artist. I'd rather have redundancy than ambguity.